requirements.txt
# Backend/FastAPI
fastapi==0.115.12
uvicorn==0.21.1
sqlalchemy==2.0.41
python-jose==3.3.0
passlib==1.7.4

# Bots
python-telegram-bot==22.1
python-dotenv==1.0.0
pyyaml==6.0

# Desarrollo (opcionales)
pylint==2.17.2
autopep8==2.0.1
pytest==7.3.1
